Retry version v1

+++ flyGame.h
@@ -76,7 +76,7 @@ public:
 	void resetAll();
 	void printScore(int score);
 	void printFilled();
-	int displayTimer();
+	float displayTimer();
 };
 
 #endif
